+++ modules/sde-submodules/single-level-bom-as-built/src/main/java/org/eclipse/tractusx/sde/submodels/slbab/SingleLevelBomAsBuiltV100.java
@@ -38,7 +38,8 @@ public class SingleLevelBomAsBuiltV100 extends SubmodelExtension {
 
 		String resource = "single-level-bom-as-built-v1.0.0.json";
 		// this is the path within the jar file
-		InputStream input = this.getClass().getResourceAsStream("/resources/" + resource);
+		//InputStream input = this.getClass().getResourceAsStream("/resources/" + resource);
+		InputStream input = SingleLevelBomAsBuiltV100.class.getResourceAsStream("/resources/" + resource);
 		if (input == null) {
 			// this is how we load file within editor (eg eclipse)
 			input = this.getClass().getClassLoader().getResourceAsStream(resource);